Add editions backend. Refactor editions defaults.
