+++ projects/Mallard/src/helpers/settings/defaults.ts
@@ -2,6 +2,7 @@ import remoteConfig from '@react-native-firebase/remote-config'
 import { Platform } from 'react-native'
 import { locale } from '../locale'
 import { Settings } from '../settings'
+import { editions } from '../../../../Apps/common/src/index'
 
 /*
 Default settings.
@@ -48,14 +49,6 @@ export const backends = [
     value: string
     preview: boolean
 }[]
-
-export const editions = {
-    daily: 'daily-edition',
-    ausWeekly: 'australian-edition',
-    usWeekly: 'american-edition',
-    dummy: 'the-dummy-edition',
-    training: 'training-edition',
-}
 
 const localeToEdition = new Map<string, string>()
 localeToEdition.set('en_AU', editions.ausWeekly)
