+++ projects/backend/package.json
@@ -19,8 +19,8 @@
         "typescript": "^3.5.3"
     },
     "scripts": {
-        "build": "ncc build index.ts -o dist -m -e aws-sdk",
-        "start": "ts-node-dev --ignore-watch node_modules index.ts",
+        "build": "ncc build backend.ts -o dist -m -e aws-sdk",
+        "start": "ts-node-dev --ignore-watch node_modules backend.ts",
         "preview": "publicationStage=preview yarn start",
         "validate": "cd ../.. && script/make validate-backend",
         "fix": "cd ../.. && script/make fix-backend",
